Created recipe to install libpng version 1.6.18. Inside we check against version 1.6.16 because that is what comes with Fedora 22. Fedora 21 is about 1 month away from EOL at the time and Fedora 23 has 1.16.17 and Rawhide has this version. Ubuntu has an ancient version.
As far as I have tested, installing this lib with pybombs has no negative side-effects on gnuradio.

I have created this recipe for gr-fosphor as it needs a newer libpng due to the use of some functions.
Also at the time of this writing gr-fosphor is not propperly linked against libpng so although it compiles, it will not run.